docs: align the OpenAPI spec with the real implementation, drop dead JSDoc, and fix three runtime issues (#11)
- Strip password/deleted_at from the PATCH /users/me response
- Cast count(*) to int in getQuestionStats
- Validate the :id param on /games/:id/config (400 instead of 500)
- Correct error messages, 429 cases, cursor errors and the missing 404 on POST /games
- Document the real GameConfigPatch filtering and normalizeConfig behaviour
- Remove the dead @openapi JSDoc blocks from the five route files

+++ backend/src/modules/game/game.repository.ts
@@ -180,8 +180,8 @@ export const getQuestionStats = async (gameSessionId: number) => {
   const { rows } = await pool.query<QuestionStatRow>(
     `SELECT
         (ans->>'question_id')::int              AS question_id,
-        count(*)                                AS answer_count,
-        count(*) FILTER (WHERE (ans->>'is_correct')::boolean) AS correct_count
+        count(*)::int                           AS answer_count,
+        (count(*) FILTER (WHERE (ans->>'is_correct')::boolean))::int AS correct_count
      FROM player_sessions ps,
           jsonb_array_elements(ps.answered_questions) AS ans
      WHERE ps.game_session_id = $1 AND ps.deleted_at IS NULL
